elevenlabs-tts/requirements.txt
elevenlabs==2.23.0
python-dotenv==1.0.0
